(PC-17806) refactor(auth): move fixture in fixtures folder
