+++ src/fixtures/user.ts
@@ -7,7 +7,7 @@ import {
   UserRole,
   YoungStatusType,
 } from 'api/gen'
-import { EIGHTEEN_AGE_DATE } from 'features/auth/pages/signup/SetBirthday/utils/fixtures'
+import { EIGHTEEN_AGE_DATE } from 'features/auth/fixtures/fixtures'
 
 export const nonBeneficiaryUser: UserProfileResponse = {
   bookedOffers: {},
