+++ src/ui/components/inputs/DateInput/DatePicker/DatePickerDropDown.web.test.tsx
@@ -1,11 +1,7 @@
 import mockdate from 'mockdate'
 import React from 'react'
 
-import {
-  MINIMUM_DATE,
-  CURRENT_DATE,
-  DEFAULT_SELECTED_DATE,
-} from 'features/auth/pages/signup/SetBirthday/utils/fixtures'
+import { MINIMUM_DATE, CURRENT_DATE, DEFAULT_SELECTED_DATE } from 'features/auth/fixtures/fixtures'
 import { fireEvent, render } from 'tests/utils/web'
 import { DatePickerDropDown } from 'ui/components/inputs/DateInput/DatePicker/DatePickerDropDown.web'
